gpt-4o-streamlit-chatbot-main/requirements.txt
openai==1.30.1 #no need
streamlit==1.34.0
pip install streamlit==1.34.0 protobuf<5 google-generativeai
google-generativeai==0.3.0
protobuf<5
google-auth==2.28.1
google-auth-oauthlib==0.7.1     